requirements.txt
Flask
MLB-StatsAPI 
requests-cache
APScheduler
firebase-admin 